# [PERF] MimeDescriptorImpl: avoid counting lines for non text parts