This allows buffering parts reads, which is used to discover its size.

Local benchmark showed
 - A good 10x speed enhancement on 650KB email (2.9ms to 0.25 ms)
 - A decent 2x speed enhancement on 5KB email (40 us to 20 us)

Context: Line length is only needed for textual parts...